Add JS timer
We need a timer to drive events outside of web audio that stays in sync with
time inside web audio. This commit has a first attempt at that.
It uses the audio-context-timers package for a stable, but inaccurate interval
that is corrected by calculating drift from the audio context currentTime.
While this does not give us events perfectly synced to the web audio clock, it
will hopefully be enough for UI updates and network requests.